Rename module for fork to point to ringerc/gripmock
I'm happy to extract these changes and send them upstream,
but in the meantime I'm editing the go module files in
my fork to point to the fork rather than upstream.

This change also removes the example go.mod and go.sum, since they
should have a separate go.mod for each client instead and that's
trivially generated on-demand.

The README is updated to point to this repo, and the example is modified
to use grpcurl instead of the sample client.